+++ server/src/main/java/org/opensearch/repositories/blobstore/BlobStoreRepository.java
@@ -1145,8 +1145,7 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
                                     // see https://github.com/opensearch-project/OpenSearch/issues/8469
                                     new RemoteSegmentStoreDirectoryFactory(
                                         remoteStoreLockManagerFactory.getRepositoriesService(),
-                                        threadPool,
-                                        recoverySettings
+                                        threadPool
                                     ).newDirectory(
                                         remoteStoreRepoForIndex,
                                         indexUUID,
@@ -1616,8 +1615,7 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
                                             // see https://github.com/opensearch-project/OpenSearch/issues/8469
                                             new RemoteSegmentStoreDirectoryFactory(
                                                 remoteStoreLockManagerFactory.getRepositoriesService(),
-                                                threadPool,
-                                                recoverySettings
+                                                threadPool
                                             ).newDirectory(
                                                 remoteStoreRepoForIndex,
                                                 indexUUID,
@@ -2550,7 +2548,7 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
      * the next version number from when the index blob was written.  Each individual index-N blob is
      * only written once and never overwritten.  The highest numbered index-N blob is the latest one
      * that contains the current snapshots in the repository.
-     *
+     * <p>
      * Package private for testing
      */
     long latestIndexBlobId() throws IOException {