+++ server/src/main/java/org/opensearch/search/slice/DocValuesSliceQuery.java
@@ -49,7 +49,7 @@ import java.io.IOException;
 
 /**
  * A {@link SliceQuery} that uses the numeric doc values of a field to do the slicing.
- *
+ * <p>
  * <b>NOTE</b>: With deterministic field values this query can be used across different readers safely.
  * If updates are accepted on the field you must ensure that the same reader is used for all `slice` queries.
  *