+++ server/src/test/java/org/opensearch/index/store/RemoteSegmentStoreDirectoryTests.java
@@ -25,9 +25,7 @@ import org.apache.lucene.util.Version;
 import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.common.UUIDs;
 import org.opensearch.common.blobstore.AsyncMultiStreamBlobContainer;
-import org.opensearch.common.blobstore.stream.read.ReadContext;
 import org.opensearch.common.blobstore.stream.write.WriteContext;
-import org.opensearch.common.io.InputStreamContainer;
 import org.opensearch.common.io.VersionedCodecStreamWrapper;
 import org.opensearch.common.io.stream.BytesStreamOutput;
 import org.opensearch.common.lucene.store.ByteArrayIndexInput;
@@ -42,7 +40,6 @@ import org.opensearch.index.shard.IndexShardTestCase;
 import org.opensearch.index.store.lockmanager.RemoteStoreMetadataLockManager;
 import org.opensearch.index.store.remote.metadata.RemoteSegmentMetadata;
 import org.opensearch.index.store.remote.metadata.RemoteSegmentMetadataHandler;
-import org.opensearch.indices.recovery.DefaultRecoverySettings;
 import org.opensearch.indices.replication.common.ReplicationType;
 import org.opensearch.threadpool.ThreadPool;
 import org.junit.After;
@@ -51,27 +48,24 @@ import org.junit.Before;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.file.NoSuchFileException;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.function.UnaryOperator;
 
 import org.mockito.Mockito;
 
 import static org.opensearch.index.store.RemoteSegmentStoreDirectory.METADATA_FILES_TO_FETCH;
+import static org.opensearch.index.store.RemoteSegmentStoreDirectory.MetadataFilenameUtils.SEPARATOR;
 import static org.opensearch.test.RemoteStoreTestUtils.createMetadataFileBytes;
 import static org.opensearch.test.RemoteStoreTestUtils.getDummyMetadata;
 import static org.hamcrest.CoreMatchers.is;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.contains;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
@@ -149,14 +143,12 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
             remoteMetadataDirectory,
             mdLockManager,
             threadPool,
-            indexShard.shardId(),
-            DefaultRecoverySettings.INSTANCE
+            indexShard.shardId()
         );
         try (Store store = indexShard.store()) {
             segmentInfos = store.readLastCommittedSegmentsInfo();
         }
 
-        when(remoteDataDirectory.getDownloadRateLimiter()).thenReturn(UnaryOperator.identity());
         when(threadPool.executor(ThreadPool.Names.REMOTE_PURGE)).thenReturn(executorService);
         when(threadPool.executor(ThreadPool.Names.REMOTE_RECOVERY)).thenReturn(executorService);
     }
@@ -213,9 +205,7 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
     }
 
     public void testGetPrimaryTermGenerationUuid() {
-        String[] filenameTokens = "abc__9223372036854775795__9223372036854775784__uuid_xyz".split(
-            RemoteSegmentStoreDirectory.MetadataFilenameUtils.SEPARATOR
-        );
+        String[] filenameTokens = "abc__9223372036854775795__9223372036854775784__uuid_xyz".split(SEPARATOR);
         assertEquals(12, RemoteSegmentStoreDirectory.MetadataFilenameUtils.getPrimaryTerm(filenameTokens));
         assertEquals(23, RemoteSegmentStoreDirectory.MetadataFilenameUtils.getGeneration(filenameTokens));
     }
@@ -392,7 +382,7 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
         remoteSegmentStoreDirectory.init();
 
         IndexInput indexInput = mock(IndexInput.class);
-        when(remoteDataDirectory.openInput(startsWith("_0.si"), eq(IOContext.DEFAULT))).thenReturn(indexInput);
+        when(remoteDataDirectory.openInput(startsWith("_0.si"), anyLong(), eq(IOContext.DEFAULT))).thenReturn(indexInput);
 
         assertEquals(indexInput, remoteSegmentStoreDirectory.openInput("_0.si", IOContext.DEFAULT));
     }
@@ -405,7 +395,7 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
         populateMetadata();
         remoteSegmentStoreDirectory.init();
 
-        when(remoteDataDirectory.openInput(startsWith("_0.si"), eq(IOContext.DEFAULT))).thenThrow(new IOException("Error"));
+        when(remoteDataDirectory.openInput(startsWith("_0.si"), anyLong(), eq(IOContext.DEFAULT))).thenThrow(new IOException("Error"));
 
         assertThrows(IOException.class, () -> remoteSegmentStoreDirectory.openInput("_0.si", IOContext.DEFAULT));
     }
@@ -568,118 +558,6 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
         storeDirectory.close();
     }
 
-    public void testCopyFilesToMultipart() throws Exception {
-        String filename = "_0.cfe";
-        populateMetadata();
-        remoteSegmentStoreDirectory.init();
-
-        Directory storeDirectory = mock(Directory.class);
-        AsyncMultiStreamBlobContainer blobContainer = mock(AsyncMultiStreamBlobContainer.class);
-        when(remoteDataDirectory.getBlobContainer()).thenReturn(blobContainer);
-
-        Mockito.doAnswer(invocation -> {
-            ActionListener<ReadContext> completionListener = invocation.getArgument(1);
-            final CompletableFuture<InputStreamContainer> future = new CompletableFuture<>();
-            future.complete(new InputStreamContainer(new ByteArrayInputStream(new byte[] { 42 }), 0, 1));
-            completionListener.onResponse(new ReadContext(1, List.of(() -> future), ""));
-            return null;
-        }).when(blobContainer).readBlobAsync(any(), any());
-
-        CountDownLatch downloadLatch = new CountDownLatch(1);
-        ActionListener<String> completionListener = new ActionListener<>() {
-            @Override
-            public void onResponse(String unused) {
-                downloadLatch.countDown();
-            }
-
-            @Override
-            public void onFailure(Exception e) {}
-        };
-        Path path = createTempDir();
-        DirectoryFileTransferTracker directoryFileTransferTracker = new DirectoryFileTransferTracker();
-        long sourceFileLengthInBytes = remoteSegmentStoreDirectory.fileLength(filename);
-        remoteSegmentStoreDirectory.copyTo(filename, storeDirectory, path, directoryFileTransferTracker, completionListener);
-        assertTrue(downloadLatch.await(5000, TimeUnit.SECONDS));
-        verify(blobContainer, times(1)).readBlobAsync(contains(filename), any());
-        verify(storeDirectory, times(0)).copyFrom(any(), any(), any(), any());
-
-        // Verify stats are updated to DirectoryFileTransferTracker
-        assertEquals(sourceFileLengthInBytes, directoryFileTransferTracker.getTransferredBytesSucceeded());
-    }
-
-    public void testCopyFilesTo() throws Exception {
-        String filename = "_0.cfe";
-        populateMetadata();
-        remoteSegmentStoreDirectory.init();
-
-        Directory storeDirectory = mock(Directory.class);
-        CountDownLatch downloadLatch = new CountDownLatch(1);
-        ActionListener<String> completionListener = new ActionListener<>() {
-            @Override
-            public void onResponse(String unused) {
-                downloadLatch.countDown();
-            }
-
-            @Override
-            public void onFailure(Exception e) {}
-        };
-        Path path = createTempDir();
-        remoteSegmentStoreDirectory.copyTo(filename, storeDirectory, path, new DirectoryFileTransferTracker(), completionListener);
-        assertTrue(downloadLatch.await(5000, TimeUnit.MILLISECONDS));
-        verify(storeDirectory, times(1)).copyFrom(any(), eq(filename), eq(filename), eq(IOContext.DEFAULT));
-    }
-
-    public void testCopyFilesToEmptyPath() throws Exception {
-        String filename = "_0.cfe";
-        populateMetadata();
-        remoteSegmentStoreDirectory.init();
-
-        Directory storeDirectory = mock(Directory.class);
-        AsyncMultiStreamBlobContainer blobContainer = mock(AsyncMultiStreamBlobContainer.class);
-        when(remoteDataDirectory.getBlobContainer()).thenReturn(blobContainer);
-
-        CountDownLatch downloadLatch = new CountDownLatch(1);
-        ActionListener<String> completionListener = new ActionListener<>() {
-            @Override
-            public void onResponse(String unused) {
-                downloadLatch.countDown();
-            }
-
-            @Override
-            public void onFailure(Exception e) {}
-        };
-        remoteSegmentStoreDirectory.copyTo(filename, storeDirectory, null, new DirectoryFileTransferTracker(), completionListener);
-        assertTrue(downloadLatch.await(5000, TimeUnit.MILLISECONDS));
-        verify(storeDirectory, times(1)).copyFrom(any(), eq(filename), eq(filename), eq(IOContext.DEFAULT));
-    }
-
-    public void testCopyFilesToException() throws Exception {
-        String filename = "_0.cfe";
-        populateMetadata();
-        remoteSegmentStoreDirectory.init();
-
-        Directory storeDirectory = mock(Directory.class);
-        Mockito.doThrow(new IOException())
-            .when(storeDirectory)
-            .copyFrom(any(Directory.class), anyString(), anyString(), any(IOContext.class));
-        CountDownLatch downloadLatch = new CountDownLatch(1);
-        ActionListener<String> completionListener = new ActionListener<>() {
-            @Override
-            public void onResponse(String unused) {
-
-            }
-
-            @Override
-            public void onFailure(Exception e) {
-                downloadLatch.countDown();
-            }
-        };
-        Path path = createTempDir();
-        remoteSegmentStoreDirectory.copyTo(filename, storeDirectory, path, new DirectoryFileTransferTracker(), completionListener);
-        assertTrue(downloadLatch.await(5000, TimeUnit.MILLISECONDS));
-        verify(storeDirectory, times(1)).copyFrom(any(), eq(filename), eq(filename), eq(IOContext.DEFAULT));
-    }
-
     public void testCopyFilesFromMultipartIOException() throws Exception {
         String filename = "_100.si";
         AsyncMultiStreamBlobContainer blobContainer = mock(AsyncMultiStreamBlobContainer.class);
@@ -689,8 +567,7 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
             remoteMetadataDirectory,
             mdLockManager,
             threadPool,
-            indexShard.shardId(),
-            DefaultRecoverySettings.INSTANCE
+            indexShard.shardId()
         );
 
         populateMetadata();
@@ -1178,6 +1055,10 @@ public class RemoteSegmentStoreDirectoryTests extends IndexShardTestCase {
         actualList.sort(String::compareTo);
 
         assertEquals(List.of(file3, file2, file4, file6, file5, file1), actualList);
+
+        long count = file1.chars().filter(ch -> ch == SEPARATOR.charAt(0)).count();
+        // There should not be any `_` in mdFile name as it is used a separator .
+        assertEquals(14, count);
     }
 
     private static class WrapperIndexOutput extends IndexOutput {